# packages/backend/src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Matrix Flow API",
    "version": "1.0.0",
    "description": "API documentation for Matrix Flow application",
    "contact": {
      "name": "API Support",
      "url": "https://github.com/yourusername/matrix-flow"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3004",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Matrix": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "projectId": {
            "type": "integer"
          },
          "isSubMatrix": {
            "type": "boolean"
          },
          "config": {
            "type": "object"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive"
            ]
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/api/execute/matrix/{id}": {
      "post": {
        "tags": [
          "Execute"
        ],
        "summary": "Execute a matrix flow",
        "description": "Executes a matrix flow with the given input",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Matrix ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "input": {
                    "type": "object",
                    "description": "Input data for the matrix execution"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Matrix execution started successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "matrixId": {
                      "type": "string"
                    },
                    "input": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Matrix is not active"
          },
          "404": {
            "description": "Matrix not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/matrix": {
      "get": {
        "tags": [
          "Matrix"
        ],
        "summary": "Get all matrices",
        "responses": {
          "200": {
            "description": "List of all matrices",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Matrix"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Matrix"
        ],
        "summary": "Create a new matrix",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "projectId"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "projectId": {
                    "type": "integer"
                  },
                  "isSubMatrix": {
                    "type": "boolean"
                  },
                  "config": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Matrix created successfully"
          }
        }
      }
    },
    "/api/matrix/{id}": {
      "get": {
        "tags": [
          "Matrix"
        ],
        "summary": "Get a matrix by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Matrix details"
          },
          "404": {
            "description": "Matrix not found"
          }
        }
      },
      "put": {
        "tags": [
          "Matrix"
        ],
        "summary": "Update a matrix",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Matrix"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Matrix updated successfully"
          }
        }
      },
      "delete": {
        "tags": [
          "Matrix"
        ],
        "summary": "Delete a matrix",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Matrix deleted successfully"
          }
        }
      }
    },
    "/api/matrix/{id}/nodes": {
      "get": {
        "tags": [
          "Matrix"
        ],
        "summary": "Get all nodes for a matrix",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of nodes for the matrix"
          }
        }
      }
    },
    "/api/matrix/{id}/connections": {
      "get": {
        "tags": [
          "Matrix"
        ],
        "summary": "Get all connections for a matrix",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of connections for the matrix"
          }
        }
      }
    },
...
  },
...
}